requirements.txt
streamlit==1.52.1
langchain==0.3.7
langchain-openai==0.2.9
openai>=1.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
langchain-google-genai
Pillow
docx2txt
PyPDF2
